+++ src/Entity/ModuleVersion.php
@@ -7,6 +7,7 @@ namespace App\Entity;
 use App\Repository\ModuleVersionRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity(repositoryClass: ModuleVersionRepository::class)]
@@ -17,7 +18,7 @@ class ModuleVersion extends AbstractBaseEntity implements \Stringable
     #[ORM\JoinColumn(nullable: false, onDelete: 'CASCADE')]
     private Module $module;
 
-    #[ORM\Column(type: 'string', length: 255, nullable: true)]
+    #[ORM\Column(type: Types::STRING, length: 255, nullable: true)]
     private ?string $version = null;
 
     #[ORM\ManyToMany(targetEntity: Installation::class, mappedBy: 'moduleVersions')]