+++ src/Repository/GitRepoRepository.php
@@ -7,6 +7,7 @@ namespace App\Repository;
 use App\Entity\GitRepo;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\Uid\Ulid;
 
 /**
  * @extends ServiceEntityRepository<GitRepo>
@@ -70,7 +71,7 @@ class GitRepoRepository extends ServiceEntityRepository
      */
     public function findPackageVersionsPerRepoWithAdvisories(): array
     {
-        /** @var list<array{repoId: \Symfony\Component\Uid\Ulid, pvId: \Symfony\Component\Uid\Ulid}> $rows */
+        /** @var list<array{repoId: Ulid, pvId: Ulid}> $rows */
         $rows = $this->createQueryBuilder('r')
             ->select('r.id AS repoId', 'pv.id AS pvId')
             ->distinct()